requirements.txt
Django==5.0.1
djangorestframework>=3.10.0
mysqlclient==2.2.1
django-environ==0.9.0
gunicorn==21.2.0
six==1.16.0
psycopg2-binary
redis==4.4.2
python-engineio==3.14.2
# python-socketio==5.11.2
python-socketio==4.6.1
uvicorn==0.30.1
daphne==4.1.2
termcolor==2.4.0
